Custom error messages
This commit implements custom messages. The main idea here is that
developers will be able to provide custom validation helpers for any
helper function they decide to use.

In case of multi argument functions, we'll check if the optional
argument is an object or a string. If it turns out to be a string, we'll
consider it to be a custom error message.